Move exercises to separate file
